created main, config and ISR files
Created the basis for the rest of the project, including a main routine to perform the RTS loop, as well as a config and ISR file to handle their respective duties.

Also configured PBCLK3 and T2 and T3 according to CST337 lab as a proof-of-concept test. All configurations appear to be working so far in the simulator.

Also added a specific datasheet for the ADC unit to the documentation folder.